+++ samples/generators/pg/src/01_tables.sql
@@ -25,7 +25,7 @@ create sequence SEQ_PROFIL start 1000 increment 50;
   * Création de la table PROFIL
  **/
 create table PROFIL (
-	PRO_ID int8 not null,
+	PRO_ID int not null,
 	TPR_CODE varchar(3),
 	constraint PK_PROFIL primary key (PRO_ID)
 );
@@ -34,7 +34,7 @@ create table PROFIL (
   * Création de la table PROFIL_DROIT
  **/
 create table PROFIL_DROIT (
-	PRO_ID int8 not null,
+	PRO_ID int not null,
 	DRO_CODE varchar(3) not null,
 	constraint PK_PROFIL_DROIT primary key (PRO_ID,DRO_CODE)
 );
@@ -47,8 +47,8 @@ create sequence SEQ_SECTEUR start 1000 increment 50;
   * Création de la table SECTEUR
  **/
 create table SECTEUR (
-	SEC_ID int8 not null,
-	PRO_ID int8,
+	SEC_ID int not null,
+	PRO_ID int,
 	constraint PK_SECTEUR primary key (SEC_ID)
 );
 
@@ -78,17 +78,17 @@ create sequence SEQ_UTILISATEUR start 1000 increment 50;
   * Création de la table UTILISATEUR
  **/
 create table UTILISATEUR (
-	UTI_ID int8 not null,
+	UTI_ID int not null,
 	UTI_AGE numeric(20, 9),
-	PRO_ID int8,
+	PRO_ID int,
 	UTI_EMAIL varchar(50),
 	UTI_NOM varchar(3),
 	UTI_ACTIF boolean,
 	TUT_CODE varchar(3),
-	UTI_ID_PARENT int8,
+	UTI_ID_PARENT int,
 	UTI_DATE_CREATION date,
 	UTI_DATE_MODIFICATION date,
-	UTI_ID_ENFANT int8,
+	UTI_ID_ENFANT int,
 	constraint PK_UTILISATEUR primary key (UTI_ID)
 );
